Update file-system.md
I was looking for AST referring to an Asynchronous Syntax Tree and found references to Abstract Syntax Tree http://en.wikipedia.org/wiki/Abstract_syntax_tree .  From the wikipedia article "In computer science, an abstract syntax tree (AST), or just syntax tree, is a tree representation of the abstract syntactic structure of source code written in a programming language. "  Possibly there should be a link to the wikipedia page?