Adding install and troubleshooting instructions
The yo install script (v1.3.2) does not install grunt/grunt-cli/bower. New users will need the additional instructions.
The recommendation to use 'sudo' as a workaround for permissions and access errors is not fool-proof, as it can break the scaffolding 'npm install' instruction.
Instead, changing the 'node-module' directory to the user's home directory, as detailed in the linked guide, will address the permissions issues.