restore previous placement of </form>
If the Todos list is included within the form, addTodo() is also called when the X button (added in the next step) is clicked. Closing the form after the first input row, and keeping the list of entered todos out of the form so that ng-submit is only bound to one button, ensures that addTodo is only called when "Add" is clicked.